rn: Give the screen components consistent names
This helps a lot with finding things!
