+++ packages/edge-login-ui-rn/src/components/screens/OtpErrorScreen.js
@@ -3,21 +3,21 @@
 import React, { Component } from 'react'
 import { Text, View } from 'react-native'
 
-import { resetOtpReset, retryWithOtp } from '../../../actions/LoginAction.js'
-import s from '../../../common/locales/strings.js'
-import OtpBackupKeyConnector from '../../../connectors/componentConnectors/OtpBackupKeyConnector.js'
-import * as Constants from '../../../constants/index.js'
-import * as Styles from '../../../styles/index.js'
-import { type Dispatch, type RootState } from '../../../types/ReduxTypes.js'
-import { EdgeLoginQr } from '../../abSpecific/EdgeLoginQrComponent.js'
-import { OtpHeroComponent } from '../../abSpecific/OtpHeroComponent'
-import { Button } from '../../common/Button.js'
-import { Header } from '../../common/Header.js'
-import SafeAreaView from '../../common/SafeAreaViewGradient.js'
-import { StaticModal } from '../../common/StaticModal.js'
-import { DisableOtpModal } from '../../modals/DisableOtpModal.js'
-import { OtpAuthCodeModal } from '../../modals/OtpAuthCodeModal.js'
-import { connect } from '../../services/ReduxStore.js'
+import { resetOtpReset, retryWithOtp } from '../../actions/LoginAction.js'
+import s from '../../common/locales/strings.js'
+import OtpBackupKeyConnector from '../../connectors/componentConnectors/OtpBackupKeyConnector.js'
+import * as Constants from '../../constants/index.js'
+import * as Styles from '../../styles/index.js'
+import { type Dispatch, type RootState } from '../../types/ReduxTypes.js'
+import { EdgeLoginQr } from '../abSpecific/EdgeLoginQrComponent.js'
+import { OtpHeroComponent } from '../abSpecific/OtpHeroComponent'
+import { Button } from '../common/Button.js'
+import { Header } from '../common/Header.js'
+import SafeAreaView from '../common/SafeAreaViewGradient.js'
+import { StaticModal } from '../common/StaticModal.js'
+import { DisableOtpModal } from '../modals/DisableOtpModal.js'
+import { OtpAuthCodeModal } from '../modals/OtpAuthCodeModal.js'
+import { connect } from '../services/ReduxStore.js'
 
 type OwnProps = {}
 type StateProps = {
