+++ packages/edge-login-ui-rn/src/components/screens/newAccount/NewAccountUsernameScreen.js
