+++ packages/edge-login-ui-rn/src/components/screens/newAccount/NewAccountWelcomeScreen.js
@@ -4,15 +4,15 @@ import React, { Component } from 'react'
 import { View } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
-import * as Assets from '../../../assets'
+import * as Assets from '../../../assets/index.js'
 import s from '../../../common/locales/strings.js'
-import T from '../../../components/common/FormattedText.js'
 import * as Constants from '../../../constants/index.js'
 import * as Styles from '../../../styles/index.js'
 import { type Dispatch, type RootState } from '../../../types/ReduxTypes.js'
 import { scale } from '../../../util/scaling.js'
 import { ImageHeaderComponent } from '../../abSpecific/ImageHeaderComponent'
 import { Button } from '../../common/Button.js'
+import T from '../../common/FormattedText.js'
 import { HeaderBackButton } from '../../common/HeaderBackButton.js'
 import SafeAreaView from '../../common/SafeAreaView.js'
 import { connect } from '../../services/ReduxStore.js'
