+++ packages/edge-login-ui-rn/src/connectors/componentConnectors/HeaderConnectorRecoveryLogin.js
